refactor: 코드 리팩토링 1. 매직넘버 정의 2. 줄 간격 정리
