+++ src/main/java/lotto/controller/LottoController.java
@@ -1,8 +1,6 @@
 package lotto.controller;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 import lotto.domain.Lotto;
 import lotto.domain.LottoStatistics;
 import lotto.domain.Lottos;
