+++ src/main/java/lotto/domain/LottoStatistics.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.util.Statistics;
 
 public class LottoStatistics {
