+++ src/main/java/lotto/domain/Statistics.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 
 public enum Statistics {
     FIRST(6, 2000000000),
-    SECOND(7, 30000000), // 이게 보너스볼
+    SECOND(7, 30000000),
     THIRD(5, 1500000),
     FOURTH(4, 50000),
     FIFTH(3, 5000),
