+++ src/main/java/lotto/util/Util.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Util {
+
     public static List<Integer> stringToIntegerList(String unrefinedNumber) {
         return Arrays.asList(unrefinedNumber.split(", ")).stream().map(Integer::parseInt).collect(
             Collectors.toList());
